🚚  Move convertion funcs to specific file
