+++ include/travesim_adapters/ros/vision_receiver.hpp
@@ -125,73 +125,6 @@ class VisionReceiver {
          * @return false Othwerwise
          */
         bool get_received_first_message();
-
-        /**
-         * @brief Static function to convert geometry_msgs::Point to travesim::Vector2D
-         *
-         * @param point Data to be converted
-         * @return travesim::Vector2D Converted data
-         */
-        static travesim::Vector2D Point_to_Vector2D(geometry_msgs::Point* point);
-
-        /**
-         * @brief Static function to convert geometry_msgs::Vector3 to travesim::Vector2D
-         *
-         * @param vector3 Data to be converted
-         * @return travesim::Vector2D Converted data
-         */
-        static travesim::Vector2D Vector3_to_Vector2D(geometry_msgs::Vector3* vector3);
-
-        /**
-         * @brief Static function to convert gazebo_msgs::ModelState to travesim::EntityState
-         *
-         * @param model_state Data to be converted
-         * @return travesim::EntityState Converted data
-         */
-        static travesim::EntityState ModelState_to_EntityState(gazebo_msgs::ModelState* model_state);
-
-        /**
-         * @brief Static function to convert gazebo_msgs::ModelState to travesim::RobotState
-         *
-         * @param model_state Data to be converted
-         * @param is_yellow True if the robot is from the yellow team, false otherwise
-         * @param id Id number of the robot
-         * @return travesim::RobotState Converte data
-         */
-        static travesim::RobotState ModelState_to_RobotState(gazebo_msgs::ModelState* model_state,
-                                                             bool is_yellow = true, int id = 0);
-
-        /**
-         * @brief Static function to convert travesim::Vector2D to geometry_msgs::Point
-         *
-         * @param vector2d Data to be converted
-         * @return geometry_msgs::Point Converted data
-         */
-        static geometry_msgs::Point Vector2D_to_Point(Vector2D* vector2d);
-
-        /**
-         * @brief Static function to convert travesim::Vector2D to geometry_msgs::Vector3
-         *
-         * @param vector2d Data to be converted
-         * @return geometry_msgs::Vector3 Converted data
-         */
-        static geometry_msgs::Vector3 Vector2D_to_Vector3(Vector2D* vector2d);
-
-        /**
-         * @brief Static function to convert travesim::EntityState to gazebo_msgs::ModelState
-         *
-         * @param entity_state Data to be converted
-         * @return gazebo_msgs::ModelState Converted data
-         */
-        static gazebo_msgs::ModelState EntityState_to_ModelState(EntityState* entity_state);
-
-        /**
-         * @brief Static function to convert travesim::RobotState to gazebo_msgs::ModelState
-         *
-         * @param entity_state Data to be converted
-         * @return gazebo_msgs::ModelState Converted data
-         */
-        static gazebo_msgs::ModelState RobotState_to_ModelState(RobotState* robot_state);
 };
 }  // ros_side
 }  // travesim
